requirements.txt
# Web Framework
fastapi==0.135.3
starlette==1.0.0
uvicorn==0.29.0
pydantic==2.10.6

# LLM Integration
langchain-google-genai==2.0.5
langchain-core==0.3.28
langchain==0.3.0

# google-generativeai>=0.7.0,<0.9.0

# ADD THIS instead
google-genai==0.3.0

# Memory
mem0ai==0.1.29

# Git
GitPython==3.1.43

# Utils
python-dotenv==1.0.1

# Scanners
semgrep
checkov
